Server 분리후 @Component 등록 및 App @ComponentScan 등록
